Added MenuItems component
